refs #9 계산기 클린아키텍쳐 적용
