+++ caclulator/starter/lib/data/model/calculator.dart
@@ -1,0 +1,5 @@
+class CalculatorModel {
+  final String result;
+
+  CalculatorModel({required this.result});
+}
