+++ caclulator/starter/lib/data/soruce/local/calculator.dart
@@ -1,0 +1,8 @@
+abstract class ICalculatorLocalDataSource {
+  String get key;
+
+  Future<void> setString(String value);
+
+  Future<String?> getString();
+}
+
